Refactoring (Resource.rcにDescript.txtを追加したのと、SkillDescription系のファイルを分けた)
