+++ SpaceWars2/Resource.rc
@@ -1,3 +1,6 @@
+// TextFile
+7601	FILE	"skillDescription.txt"
+
 // Fonts
 7700	FILE	"fonts/03SmartFont-P.ttf"
 7701	FILE	"fonts/Cica-Regular.ttf"
