+++ SpaceWars2/SpaceWars2.vcxproj
@@ -173,6 +173,7 @@
     <ClCompile Include="functions\KeyRepeat.cpp" />
     <ClCompile Include="functions\MainSkill.cpp" />
     <ClCompile Include="functions\Player.cpp" />
+    <ClCompile Include="functions\SkillDescription.cpp" />
     <ClCompile Include="functions\SpecialSkill.cpp" />
     <ClCompile Include="functions\SubSkill.cpp" />
     <ClCompile Include="functions\Vec2Utils.cpp" />
@@ -314,6 +315,7 @@
     <ClInclude Include="functions\KeyRepeat.hpp" />
     <ClInclude Include="functions\MainSkill.hpp" />
     <ClInclude Include="functions\Player.hpp" />
+    <ClInclude Include="functions\SkillDescription.hpp" />
     <ClInclude Include="functions\Vec2Utils.hpp" />
     <ClInclude Include="functions\XInput.hpp" />
     <ClInclude Include="scenes\Include.hpp" />
